requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
PyYAML>=6.0
schedule>=1.2.0
